[bldr-build] Add chroot-based build environments.
chroot Environments
===================

This change introduces 3 new environments for developing bldr Plans and
producing packages. In order to guarentee there is no accidental
contamination of library linking from a base operating system, we can
build our software in a sandboxed environment thanks to the `chroot(8)`
command.

Lower level software components (think: glibc, binutils, etc.) are much
more sensitive to the environment in which they are built, and as a
result do not always behave correctly when built in a cgroup/namespaced
environment which Docker provides us. As a hedge to the pure container
world, 2 additional build environments are available for use:

* A Vagrant-based virtual machine, backed by VirtualBox and a known-good
  Bento box (Vagrant base box)
* A "bare metal" environment which can be used on physical servers, or
  other virtual machine instances such as Amazon's EC2 or Microsoft's
  Azure cloud

Unified User Experience
=======================

All 3 options (Docker container, Vagrant virtual machine, bare metal)
have the same setup and user flow--the only difference is the `make`
target name.

Docker
------

**Note:** The usual bldr/Docker environment setup is assumed here, see
the main project's README.md for more details.

To start a shell session:

```
$ cd plans/support/chroot
$ make shell
```

To only create the base image (`make shell` automatically calls this):

```
$ cd plans/support/chroot
$ make image
```

To clean up the data container:

```
$ cd plans/support/chroot
$ make clean-data
```

Vagrant
-------

**Note:** An installation of Vagrant and VirtualBox will be required for
this solution to be used. The Bento box will be downloaded automatically
from Atlas on first boot, and this may take a short while initially.

To start a shell session:

```
$ cd plans/support/chroot
$ make vm-shell
```

To only create the base vm (`make vm-shell` automatically calls this):

```
$ cd plans/support/chroot
$ make vm
```

To destroy the vm:

```
$ cd plans/support/chroot
$ make vm-destroy
```

Bare Metal
----------

**Note:** The only real requirements for this solution are a Linux-based
distribution, and the `sudo(1)`, `make(1)`, and `chroot(8)` commands in
the `$PATH`.

To start a shell session:

```
$ cd plans/support/chroot
$ make metal-shell
```

To only create the base chroot (`make metal-shell` automatically calls
this):

```
$ cd plans/support/chroot
$ make metal
```

To destroy the chroot:

```
$ cd plans/support/chroot
$ make metal-destroy
```