[bldr-build] Prefix public callback functions with `do_`.
Ths addresses an issue where the author of a Plan wants to call the
`install(1)` command to copy files, create directories, etc. Here's a
short example that happens in an `install()` callback:

```sh
install() {
  make install
  install -D -m644 LICENSE ${pkg_path}/share/LICENSE
}
```

The call to `install` on line 3 won't execute the `install(1)` command,
but will instead recursively call the defined `install()` shell
function, which recurses into foreverland.

Additionally, now all user-facing callbacks that are designed to be
overwritten are of the naming form `do_*()`. The following functions are
now part of the public API for Plan files:

* `do_begin()` (formerly `bldr_begin()`)
* `do_download()` (formerly `download()`)
* `do_verify()` (formerly `verify()`)
* `do_clean()` (formerly `clean_cache()`)
* `do_unpack()` (formerly `unpack()`)
* `do_prepare_wrapper()` (formerly `prepare_wrapper()`)
* `do_prepare()` (formerly `prepare()`)
* `do_build_wrapper()` (formerly `build_wrapper()`)
* `do_build()` (formerly `build()`)
* `do_install_wrapper()` (formerly `install_wrapper()`)
* `do_install()` (formerly `install()`)
* `do_build_config()` (formerly `config()`)
* `do_service()` (formerly `service()`)
* `do_strip()` (formerly `strip_binaries()`)
* `do_docker_image_wrapper()` (formerly `dockerfile_wrapper()`)
* `do_dockerfile_inline()` (formerly `docker_inline()`)
* `do_docker_image()` (formerly `dockerfile()`)
* `do_end()` (formerly `bldr_end()`)