[plans] Add gcc-libs plan to replace the libgcc/libstdc++ plans.
The chef/gcc-libs plan will be built in exactly the same way that the
chef/gcc plan is built, except for its install phase which only uses
the shared libraries from gcc. The consequence is that libgcc and
libstdc++ are both in the resulting package which should (hopefully) be
convenient and not a net waste in the long term.