[plans] Break Plan checks into `do_check()` phase.
As of e1cdae4ee50e84cf0a52e071fb02310b9177dc84, there is an opt-in phase
called `do_check()` which will be executed if declared and with the
`$DO_CHECK` environment variable set to a non-empty value. The pay off
is seeing a lot of `do_build()` phase becoming the defaults (and thus
can be deleted).