[bldr-build] Delay creating `$pkg_path` until `_build_environment()`.
This change should not impact any existing Plans, unless they are doing
borderline bad things.

Prior to this change the `$pkg_path` (i.e. the full path to the
extracted package on disk, such as
`/opt/bldr/pkgs/acme/software/1.2.3/201501020304`) was created almost at
the start of the Plan execution, even before the `do_begin()` phase.
This change delays creation until after dependencies are resolved and
present, the software has been downloaded, verified, and extracted.

The benefit is that any failures of the Plan before this point would not
end up with an empty `$pkg_path` directory on disk. This also helps
Plans that determine some of their metadata later than at start.