[plans] Update bash path calculation in binutils & gcc plans.
This value is used for the shebang of several wrapper scripts around
`ld`, `gcc`, etc. In the former chroot environment, `/bin` was on the
development `$PATH` and so the `command -v bash` invocation would yield
an appropriate result.

This change forgoes the clever version and sets a hardcoded value of
`/bin/bash` as it will exist in both the stage1 and bldr Studio types.
Once the second pass of these Plans is complete, there will be an
appropriate `pkg_path_for bash` call.