[plan] Update gmp to build in a bldr Studio.
Note that the `--enable-cxx` configure flag was removed as it seems that
nothing downstream will need the C++ bindings to libgmp. Adding this extra
shared library was requiring a dependency on `chef/gcc-libs` for libstdc++. It
was not possible after an hour of trying to build this library with
`-static-libstdc++` with the desired effect, so it was easier to drop this
feature support.