[bldr-build] Delay the creation of `$pkg_path` until `do_install()`.
This change is intended to lead to less empty paths under
`/opt/bldr/pkgs/` when a build fails. Prior to this change, the target
package path (known as `$pkg_path`) was created in the
`_build_environment()` command before the `do_prepare()` phase. When
looking at how a package should be built, there is no reason for this
path to exist until just before the `do_install()` phase. If a Plan
author needs this path to exist before `do_install()` then there is most
likely something they can do to change their approach, or they can
run `mkdir -p $pkg_path` explicitly themselves. A careful study of
existing Plans show that this is a safe change and leads to less false
build directories.