[bpm] Allow `bpm install` to complete if no remote pacakge is found.
Prior to this release `bpm install` would fail early if no remote
package was found and if a suitable local package was present. The
desired behavior would be to potentially warn that no remote package
existed but ultimately complete as a local package exists. That is what
this change introduces.

Worse, without this change `bldr-build` has a sub-optimal behavior when
building Plans that depend on locally built packages that have not yet
been published to a repo.