bldr-build and studio improvements
When writing the SHA256 checksums to the manifest, files that have
spaces in the names are not escaped/quoted, so errors will get
displayed and those files will not be included in the manifest. Adding
the `-print0` option to find, along with the `-0` option to xargs
resolves this problem.

We love colorized prompts. However, in most of our default terminal
settings the escape codes cause weird behavior that prevents the
terminal from displaying really long commands. For now, let's just
have a simple prompt with no color.