Write out requirements.txt in aws-cli plan
Starting to do this on all plans that use `pip`. What this does is drops
a file in the root of the package that looks like:

```
awscli==1.10.43
botocore==1.4.33
colorama==0.3.3
docutils==0.12
jmespath==0.9.0
pyasn1==0.1.9
python-dateutil==2.5.3
rsa==3.4.2
s3transfer==0.0.1
six==1.10.0
```

so if you want to see what the exact `pip` dependencies are included in
the package it's right there for you.

Also:

* Fixes to make ShellCheck happy
* Wrap lines in `pkg_description`
* Don't delete the Python/VirtualEnv symlinks in /bin. It doesn't hurt
  anything if they exist, and it's more work to delete them than is
  needed.

Signed-off-by: Nathan L Smith <smith@chef.io>