Change name of protobuf2.5 plan
The `pkg_name` of the protobuf2.5 plan must also be `protobuf2.5`, or
any built packages would go into the namespace for core/protobuf, whose
plan uses a different version.

Set a `pkg_distname` in the plan and use that to make the plan build.

Fix the protobuf-c plan to use the correct ident.

Signed-off-by: Nathan L Smith <smith@chef.io>