cruft removal and readline-rb
This plan looks a bit plagiarized from the omnibus ruby script and
carries some cruft that badly needs to get removed.

libiconv hasn't been need in ruby forever.  We attempted to remove
it from omnibus, though, and broke people on upgrade in a minor
version because `$GREAT_NOKOGIRI_SADNESS`.  Lets not start going down
that road in habitat right from the start.

Most of the flags that disable gdbm/dbm/gmp seem pretty pointless
with the clean build system.  We had those in omnibus because we
had to in order to ignore any O/S provided libraries that might
get accidentally slurped up.  This is probably the biggest reason
about why habitat is a better build system than omnibus.

Also swapped out libedit for readline-rb since we're going to want
that instead down the road.  In that case the road actually is pretty
well trod and we know how it ends.

Now that iconv and libedit are gone, its possible ncurses could die
as well, along with some other deps but I'm not certain yet...

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>