[rust,cargo-nightly] Add wrapper for `cargo` to set SSL certs.
This change adds a wrapper script for `cargo` in both the rust and
cargo-nightly plans to depend on and set a suitable set of SSL
root certificates.

Prior to this change, and plans that built Rust projects with cargo
would have to export OpenSSL's `CA_CERT_FILE` environment variable,
otherwise fetching dependencies or registry snapshots would fail. Rather
than pushing the reponsibility onto each downstream Plan, this change
"owns" that job as it is an implementation detail after all.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>