Add `PLAN_CONTEXT` variable to build order ruby scripts.
This change is to support a refactoring in the Habitat core repo to
source and share one version file across multiple components. As the
paths vary it was necessary to use the `$PLAN_CONTEXT` variable to
compute the correct directory to the `VERSION` file. This change adds
the `PLAN_CONTEXT` variable in the Plan-sourcing inlined Bash program
and will make all Plans sourcable once again.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>