Add mercurial 3.8.1
The mercurial built by this plan is capable of cloning HTTP and HTTPS
based mercurial repositories.

Depending on your studio setup, you may see the following post-test
failure:

    --- /hab/cache/src/mercurial-3.9.1/tests/test-obsolete.t
    +++ /hab/cache/src/mercurial-3.9.1/tests/test-obsolete.t.err
    @@ -972,8 +972,6 @@

       $ chmod 000 .hg/cache/hidden
       $ hg log -r . -T '{node}' --debug
    -  cannot read hidden cache
    -  error writing hidden changesets cache
       8fd96dfc63e51ed5a8af1bec18eb4b19dbf83812 (no-eol)
     #endif

    ERROR: test-obsolete.t output changed

This fails in containers if the filesystem allows the read on
`.hg/cache/hidden` despite the 000 mode.

A number of other tests are skipped, mostly related to
internationalization and canonical-specific features.

Signed-off-by: Steven Danna <steve@chef.io>