Remove pkg_prefix or bin dirs from pkg_svc_run
This change removes the pkg_prefix and/or "bin" directories from the
pkg_svc_run variable in several plans as we're not using the
pkg_prefix in the generated run script anymore (as of 0.9.3).

We also remove the PATH setting in the gnatsd plan, as GOROOT will be
empty and we have the right path during the build. In order to find
the binary, we symlink it in the bin directory.