[python] Explicitly link gcc_s
In #316 @carrete provided a test case in which python fails with the
infamous pthread_cancel error:

    # /hab/pkgs/core/python/3.5.2/20161102165726/bin/python test.py
    Thread 0 has started
    Thread 1 has started
    Thread 1 is done
    libgcc_s.so.1 must be installed for pthread_cancel to work
    Aborted

Explicitly adding `-lgcc_s` to LDFLAGS fixes the test case:

    # /hab/pkgs/core/python/3.5.2/20161205215319/bin/python test.py
    Thread 0 has started
    Thread 1 has started
    Thread 2 has started
    Thread 2 is done
    Thread 1 is done
    Thread 0 is done

Note that `make test` is currently failing with other unrelated errors.

We've used this method of linking gcc_s in the plans for apr, guile,
tcl, and tor.

Signed-off-by: Steven Danna <steve@chef.io>