Remove InSpec from core plans
InSpec's plan is maintained at https://github.com/chef/inspec and
the official packages are published to chef/inspec, not core/inspec.
To avoid any confusion, proposing we remove this from core plans as
it is out-of-date and we don't want any community members to try to
build from an incorrect plan.

Signed-off-by: Adam Leff <adam@leff.co>