Fix python scaffolding and properly support pip.
This fixes issues with the clean phase which would bomb out in most
cases. Additionally, we have added support for building packages via
`pip install`.

Under the covers this leverages venv for Python 3.6, and depends on pip
in the venv to install the applications with the correct interpreter
paths for the shebangs.

- Update aws-cli version and use python-scaffolding

This updates the version of aws-cli to the current release and
leverages scaffolding to provide a working example of how to use the
python-scaffolding.

- Improve scaffolding-python docs and fix small bug.

This fixes a small bug in the pip arguments constructor whereby
the package name and package version should not exist in the variable.

Additionally, the documentation has been expanded to reflect the new
pip features.

- Use venv for setuptools installed packages.

This breaks out the calls to setuptools into a new function and
leverages venv to ensure the correct relative PYTHONPATH is set when
running the application.

- Automagically set pkg variables from setup.py.

Added a new `scaffolding_python_prepare` callback which attempts to
set `pkg_description` and `pkg_upstream_url` if a `setup.py` is found
and the values are defined.

- Clarify runtime dep resolution for python-scaffolding.

- Port python scaffolding to python2 with virutalenv

This ports the features and functionality from scaffolding-python to
the scaffolding-python2 package. The primary difference in the
scaffolding.sh is the default usage of `virtualenv` instead of the
`venv` which is built-in to Python 3.6 and later.

- Add support for embedded plans to python-scaffolding

This adds support for embedded plans into `habitat` or the root level
of the source. Some additional refactoring was done to clean up the
scaffolding to use the new `disable_callbacks` helper function.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>