[build-base-plans.sh] Update logic for split repos usage.
This change updates the `bin/build-base-plans.sh` script to assume that
it is being invoked from a parent directory which contains git clones of
both `habitat` and `core-plans` repositories. None of the base plans nor
their order have been changed--this was the minimal amount of work to
make this program viable once more.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>