[s-ruby] Only set hash entries if not already set.
This change changes how some Scaffolding data is set when it is in an
associative array. For `scaffolding_env`, `scaffolding_process_bins`,
`pkg_binds`, and `pkg_exports` a key is first checked to see if there is
already a value (which would have been set in a Plan) and a value will
be set only if there is no current value set.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>