[s-ruby] Attempt to only add default.toml data if not set in Plan.
This change attempts to allow a Plan author to write configuration in
their `default.toml` file which will not be overridden by the
Scaffolding code. Before a Scaffolding default entry is added, the
Plan's `default.toml` is checked first to see if a key exists there. If
no key exists, the Scaffolding default is added and otherwise it is not.
This should allow a Plan to always win with their configuration data.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>