[glibc] Fix regression due to linting in #475.
(Editor's note) sometimes I really dislike fixing Plan code because
shellcheck has an issue with it--the Plans are guarenteed to run under
GNU Bash 4.x and thus Plans don't fall prey to common shell coding
issues. This was another example of a carefully made lint fix that
resulted in a broken package which took many downstream dependencies to
find. Thankfully, this is only the frustration of someone who needs to
build **all** Plans which include the most complex cases. My hope is
that almost everyone else never has to worry about these details--we do
this in service of the greater good.(/Editor's note)

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>