[s-ruby] Support external database connection via optional bind.
This change slightly alters the auto-generated service binding when
build a Ruby-based app when a PostgreSQL database requirement is
detected. Before this change a **required** binding was generated,
meaning that the PostgreSQL instance had to be running in joined Habitat
ring under a Supervisor.

Running the database under Habitat and allowing your app service to
detect and connect remains exactly the same (by adding the `--bind
database:<service_group>` option when loading and/or starting the app).

However, if your database is currently running external to a Habitat
ring, or if your database is being managed by a cloud provider (such as
RDS with Amazon Web Services), then you must provide 2 additional config
settings:

* `db.host`: The database hostname. There is no valid default.
* `db.port`: The database port. It defaults to `5432`.

It is hoped that allowing an optional database bind will provide an even
easier way to experiment with Habitat in a mixed or transitional
environment, or those where the database remains outside the control of
the application developers.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>