[socat] Upgrade to latest, add openssl and readline support
With the addition of openssl and readlines the only socat features we
don't support are libwrap and openssl-fips. We've stayed away from
libwrap so far and I think that is a good idea.  FIPS is a can of
worms I don't want to open right now.

I've also chronicled my attempt and getting the test suite passing.
It appears that our build of socat is fine and most of the issues are
environmental; however, after solving a half dozen small problems I
decided it was OK to not have the tests for now.

Signed-off-by: Steven Danna <steve@chef.io>