provide a default database config for asset compilation
While Rails core no longer requires a database configuration by default
during asset compilation, some gems assume a database is available on
load and some Rails applications have ActiveRecord calls in routes.
Either will trigger a database config validation during
assets:precompile. Setting DATABASE_URL to a connection string that
parses cleanly will appease these applications.

Opted to check for and use an existing DATABASE_URL in case a plan for
an app actually does need a connection string to a running database. The
logistics for setting all that up is outside the scope of scaffolding--
and really is a path fraught with suffering that should be avoided--but
it is easy enough to not clobber something that may have been set by the
plan author already.

Signed-off-by: Robb Kidd <robb@thekidds.org>