Removes config servers and relies on binding
I found that when launching `core/haproxy` that the binding
is required so I don't know a situation where the supervisor
would ever use the default values found in the configuration.

I found that `has_backend` to not work correctly and always
return a false value. Making it so the members found in the
backend binding to never appear in the configuration and
for haproxy to always fall back on the default values.

I replaced the loop for writing out the backends with an
`eachAlive` which is an equivalent of `each` with the check
to see if it is alive inside the loop.

Signed-off-by: Franklin Webber <franklin@chef.io>