move setting scaffolding prefix until after pkg_version is computed
scaffolding_app_prefix isn't truly needed until do_prepare or later.
Moved setting it into scaffold's do_default_prepare.

Before it was being set suuuuper early in the scaffolding build. Because
it uses $pkg_prefix, when set so early, it did not benefit from
$pkg_prefix getting updated by update_pkg_version() which would leave
some paths to things with a version of __pkg__version__unset__.

Instead, at the early point where it used to be set, $app_prefix is set
to denote the directory inside the package into which the application
will be copied.

For setting pkg_bin_dirs, $app_prefix ("app") is used instead of
computing the basename of $scaffolding_app_prefix (also just "app").

For pkg_svc_run, because the bin/ directory has already been added to
pkg_bin_dirs and is therefore on the runtime PATH, only the executable
needs to be named there, not the full path to it.

Signed-off-by: Robb Kidd <robb@thekidds.org>