Initial xorg spike
This PR adds the packages required to build the xeyes test program as
well R and cairo (to be updated in a follow-up PR) with xlib
support. The original plan was to build the entire stack in the xorg/
origin until we had a more complete distribution. However, this was
making it difficult to incrementally add X support to applications in
core. The new plan is to let the needs of other applications in core
guide what we package from Xorg.

Notable decisions:

- All package names have been downcased for consistency

- libX11 is named xlib to make it easier for users who don't know
  about X's package structure and "just want to compile with xlib
  support"

I've added an XORG_README.md and xorg_build_order to aid with future
Xorg packaging work.

Signed-off-by: Steven Danna <steve@chef.io>