Include tools to investigate hardened builds
The desire for more robust default build flags was raised in
habitat-sh/core-plans#587. The goal of this change is to add a couple
of tools to make it easier to start investigating hardening
core-plans:

 - a shell script that can be included in plans to enable
   compiler- and linker-level security features.

 - a new bin/check-security-feature.sh script can be used
   to check that the resulting package actually has the features in the
   compiled executables.

Currently, we include the following flags:

- `-fstack-protector-strong`
- `-Wl,-z,relro`
- `-Wl,-z,now`
- `-D_FORTIFY_SOURCE=2 -O2`

For details on these flags, see the in-source comments.

* Possible Future Work

I've excluded PIE-related flags for now because I've found them hard
to set generally via CFLAGS. It may be easier to pursue PIE builds
with a gcc built with `--default-enable-pie`.

Another often-used set of flags are format-security warning flags.
These flags would be useful if and only if we have a plan to maintain
patches to fix the warnings or contribute fixes upstream.

For information on what other large distributions do:

- https://wiki.debian.org/Hardening
- https://wiki.gentoo.org/wiki/Project:Hardened
- https://fedoraproject.org/wiki/Changes/Harden_All_Packages

Signed-off-by: Steven Danna <steve@chef.io>