Fix postgresql health_check when alternate port is specified
We need to specify the configured port for pg_isready. This patch also
adds the configured superuser which will quiet the postgresql log and
prevent each health check from emitting a log entry.

The health check will now return CRITICAL if the checker is unable to
connect as well as if the connection is rejected (a not yet ready
state). While transient, this answers more clearly the question of
health from the perspective of a service depending on postgresql,
namely, "can you accept my connection and answer my queries?".

Signed-off-by: Seth Falcon <seth@chef.io>