make git a build dep in dynamic language scaffolds
This changes the git detection from checking for a `.git/` directory in
only the `SRC_PATH` to using git itself to determine whether the current
working directory exists within a git work tree.

An example where looking for `.git/` in `SRC_PATH` returns a false
negative is git repositories with multiple components in
sub-directories. The studio is generally entered at the root of the git
repository and the build command executed with a sub-directory path. The
fallback to `find | tar` ends up including many unnecessary or undesired
files in the resultant package.

So long as the studio is entered from a location that includes the git
repo root, this new git detection will be correct and the `git ls-files`
used to collect the file list for packaging will be relative to the
`SRC_PATH`.

* `_add_git` - a new function that injects `core/git` as a build
dependency during scaffolding load; takes the place of the call to old
`_detect_git`

* `_detect_git` - modified to perform git work tree detection with the
git command itself; moved calling this function to do_default_prepare
which is executed after dependencies have been retrieved and before any
portion of the scaffolding build/install needs to know about whether the
build is occurring in a git repo

Signed-off-by: Robb Kidd <robb@thekidds.org>