add plans for all currently-maintained upstream ruby versions
These plans source ruby/plan.sh because so far it looks like the steps
to build Ruby are the same for all these versions.

update_pkg_version is used to handle updating attributes to the
different version from the latest Ruby. Some attributes aren't
recomputed there, though--like pkg_source and pkg_dirname (because it's
set to _something_ already)--so they are set again in the specific
version plans.

The specific plan that tracks the latest version--ruby24 here--is a
simple sourcing of the main ruby plan with a sanity check that the
pkg_version already set is the X.Y expected. This is intended to
safeguard against the ruby plan moving to X.Z and having the specific
version get built and have its name be a lie. e.g. Upgrading ruby to
2.5.0 and having ruby24 get built with that newer version.

Signed-off-by: Robb Kidd <robb@thekidds.org>