upgrade curl to 7.58.0 to address CVE-2018-1000007
https://nvd.nist.gov/vuln/detail/CVE-2018-1000007
https://curl.haxx.se/docs/adv_2018-b3bf.html

Exerpts from curl's write-up:

- VULNERABILITY

curl might leak authentication data to third parties.

When asked to send custom headers in its HTTP requests, curl will send
that set of headers first to the host in the initial URL but also, if
asked to follow redirects and a 30X HTTP response code is returned, to
the host mentioned in URL in the Location: response header value.

Sending the same set of headers to subsequest hosts is in particular a
problem for applications that pass on custom Authorization: headers, as
this header often contains privacy sensitive information or data that
could allow others to impersonate the curl-using client's request.

- THE SOLUTION

In curl version 7.58.0, custom Authorization: headers will be limited
the same way other such headers is controlled within curl: they will
only be sent to the host used in the original URL unless curl is told
that it is ok to pass on to others using the
CURLOPT_UNRESTRICTED_AUTH[1] option.

[1] https://curl.haxx.se/libcurl/c/CURLOPT_UNRESTRICTED_AUTH.html

NOTE: this solution creates a slight change in behavior. Users who
actually want to pass on the header to other hosts now need to give curl
that specific permission. You do this with --location-trusted[2] with the
curl command line tool.

[2] https://curl.haxx.se/docs/manpage.html#--location-trusted

Signed-off-by: Robb Kidd <robb@thekidds.org>