[dep] fix build version, bump version
Our current `core/dep` package happens to be version 0.3.0 (I believe!),
however, rebuilding it would give us the latest version.

This change of plan.sh adds an explicit reset to the desired version of
dep to the download function.

Using the upstream-provided build script instead of the scaffolding lets
us build a dep binary that comes with proper version output:

    [18][default:/src:0]# dep version
    dep:
     version     : v0.4.1
     build date  : 2018-02-12
     git hash    : 37d9ea0a
     go version  : go1.9.4
     go compiler : gc
     platform    : linux/amd64

Also, this commit updates the version to the latest release of dep.

Signed-off-by: Stephan Renatus <srenatus@chef.io>