elasticsearch 6.2.x: use coreutils instead of busybox
ES 6.2.x creates a tempdir on startup using the `mktemp` utility.
The logic in the ES start script determines the flags to use for
`mktemp` based on platform and version of `mktemp`, but does not
take into account the BusyBox version at all.

See:
* elasticsearch/elasticsearch#27609
* elasticsearch/elasticsearch#27659

This results in the following errors when trying to start ES with
BusyBox utils as oppsoed to coreutils:

```
elasticsearch.default(O): mktemp: Invalid argument
```

Signed-off-by: Stephen Delano <stephen@chef.io>