[cacerts] Update Plan to use `pkg_version()`.
Historical note: this was the first Plan to use a dynamically set
version based on the downloaded source. It's needed to be occasionally
updated as it needed to rewrite various build variables as the build
program evolved, but now we have the `pkg_version()` and
`update_pkg_version()` helpers so this change uses the officially
supported mechanism. Yay!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>