[build-base-plans] Remove `libidn` from base plans.
This package is no longer in the critical packages which form the "base"
set (i.e. the minimal set of packages required to bootstrap Habitat on
Linux).

Formerly this was used in `wget` for Internationalized Domain Names
(IDN) support, but now appears to require a newer version of libidn and
an additional library.

Until this support is re-added, we don't need this package in the base
set.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>