[infra] Extra base Plans to static file & rename to build-plans.sh.
This change extracts the pure data list of the base Plans to
`base-plans.txt` in the root of the project and updates the formerly
known `bin/build-base-plans.sh` to the more generalized
`bin/build-plans.txt`. The `build-plans.sh` program will now take a list
of directories with Plans either as a file argument or on standard in
and build them in the same build-resuming way as the former script,
except now for any set of packages.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>