[infra] Add specific scripts to simplify bootstrapping.
This change adds some scripts that simplify the instructions to follow
in the internals/bootstrapping docs page. Previously, these scripts
where documentation instructions where the user had to set environment
variables, enter a Studio, set more environment variables, etc. which
proved to be error prone and needlessly complicated. These scripts aim
to simplify the bootstrapping narrative (i.e. the broad steps to follow
in order) and leave the details to (hopefully) straight-forward small
wrapping scripts which can be inspected to learn the particulars.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>