Fix: apxs requires a Perl interpreter
The `apxs` tool needs a Perl interpreter. When the `httpd` package
is built, if no Perl interpreter is found my autotools, there will
be an unrendered template Perl shebang at the top of the apxs tool.

Adding `core/perl` as a dependency fixes both the shebang, and
ensures that core functionality expected by `httpd` package users
is available.