Add Envoy plan
Let me start out with apologizing for how gross this plan is.
Envoy doesn't publish binaries that I could find and the Bazel build proved too difficult for my mere mortal abilities.
So, that led me to download the dockerfile they publish to extract the envoy binary, then patch it.

Signed-off-by: Elliott Davis <elliott@excellent.io>