Initial commit of plan variants
This is the set of base plans that require variations to work under the "x86-64-linux-kernel2" package target introduced in habitat-sh/core#42.
These plans are specifically locked to older versions of software in order to ensure compatibility with Linux kernel versions as low as 2.6.32.

There will be no differences from how you build and run Habitat plans and packages today. As we build out packages for the "kernel2 variant" you will see a hab cli with an Active Target of 'x86-64-linux-kernel2'. This will allow you to build and consume packages for the "kernel2 variant" if your environment requires support for Linux systems running on that era of the kernel.

As this is foundational for work that is still in flight, we don't have any new blog posts or documentation to share at this time. Please look forward to that as this work nears a more complete state. You can see some of the early discussion here: habitat-sh/core-plans#1652 and the initial blog post here: https://www.habitat.sh/blog/2018/06/core-plans-and-certain-linux-kernels/

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>