ensure bundler version is the version number
This commit ensures that the `_bundler_version` variable is set only
to the version as returned by the `bundle --version` command. Bundler
itself does not differentiate error and warning messages to STDERR, so
if there are problems such as a non-extistent $HOME directory, those
messages aren't partially captured in this output.

See habitat-sh/habitat#5569 for background.

Signed-off-by: Joshua Timberman <joshua@chef.io>