[postgresql] Only chown data directories we create
If you mount the data volume for postgres (e.g. via k8s) there will be
the usual `lost+found` directory owned by root.  We were trying to chown
it, failing and then postgres wouldn't start.

This PR changes the permissions setting logic in the init hook to only
cheange the permissions on the 2 data directories we create (`pgdata`
and `archive`)

Signed-off-by: James Casey <james@chef.io>