Update tar 1.30 -> 1.31
Remove the patch that's not necessary with 1.31 as well.

   tar:
   tar: Source Path: /hab/cache/src/tar-1.31
   tar: Installed Path: /hab/pkgs/tas50/tar/1.31/20190201001335
   tar: Artifact: /src/tar/results/tas50-tar-1.31-20190201001335-x86_64-linux.hart
   tar: Build Report: /src/tar/results/last_build.env
   tar: SHA256 Checksum: 5a9ef5c18d8fe298dc6905c69bf74f51901adf7ba20d257c130e55efd4e98fd9
   tar: Blake2b Checksum: f17e10c3fe9d9103f14dce19e68c152cf85c616cfb0710875816bcf0f19e486c
   tar:
   tar: I love it when a plan.sh comes together.
   tar:
   tar: Build time: 0m53s

[22][default:/src/tar:0]# /hab/pkgs/tas50/tar/1.31/20190201001335/bin/tar --help
Usage: tar [OPTION...] [FILE]...
GNU 'tar' saves many files together into a single tape or disk archive, and can
restore individual files from the archive.

List of changes in this release:

* Fix heap-buffer-overrun with --one-top-level.

Bug introduced with the addition of that option in 1.28.

* Support for zstd compression

New option '--zstd' instructs tar to use zstd as compression program.
When listing, extractng and comparing, zstd compressed archives are
recognized automatically.
When '-a' option is in effect, zstd compression is selected if the
destination archive name ends in '.zst' or '.tzst'.

* The -K option interacts properly with member names given in the command line

Names of members to extract can be specified along with the "-K NAME"
option. In this case, tar will extract NAME and those of named members
that appear in the archive after it, which is consistent with the
semantics of the option.

* Fix CVE-2018-20482

When creating archives with the --sparse option, previous versions of
tar would loop endlessly if a sparse file had been truncated while
being archived.

Signed-off-by: Tim Smith <tsmith@chef.io>