copy data_bags directory
As a Chef user, I have some data stored in JSON files that I would
like to use in Chef recipes as Data Bag Items. In order to facilitate
this, one can create a "data_bags" directory and store the bags
(directories) and files (.json within those bags), and then use the
scaffolding with this update to copy them over in the package and
they'll be available in recipes using the Chef DSL's `data_bag` and
`data_bag_item` methods.

The data bags and policyfiles directories are configurable using a
`scaffold` variable, during the `scaffolding_load` function.

* `scaffold_policyfiles_path`
* `scaffold_data_bags_path`

These default to `policyfiles` and `data_bags` in the `$PLAN_CONTEXT`
parent.

I didn't test this using encrypted data bags, though I presume it
would work given a recipe written to make use of a secret dropped off
in a known location.

Signed-off-by: Joshua Timberman <joshua@chef.io>