Replace dep on PG server client in ruby scaffolding
Ruby applications that connect to PostgreSQL databases only need to
depend on the postgres client tools. When the scaffolding assembles the
list of `pkg_deps` for a ruby project, it was adding `core/postgresql`
to the list if it detected the `pg` gem, which adds the entire server to
the installation footprint on disk (or your docker images).

This change replaces the dependency with `core/postgresql-client`, which
is smaller in size by > 100MB and ensures we only depend on exactly what
we need.

Closes #2595

Signed-off-by: Keith Walters <keith.walters@cattywamp.us>