[docs] remove package dir existence test
As discussed in https://github.com/habitat-sh/core-plans/pull/2604

Testing that the package directory exists can be useful, to provide some context when a test run fails.
However it is not really a test of the core plan, it's a test that the setup is valid.
Having this test in a plan's test suite is acceptable (imo) but to keep the standard as neat as possible, let's remove it from the 7zip example.

Signed-off-by: James Stocks <jstocks@chef.io>