check-version.sh: a script to help get us up-to-date
This script uses the release-monitoring.org API to poll the current
versions of software and compare it to the version currently in the
plan.

It uses an annotation in the plan itself to line up the plan with a
project in the release-monitoring API.

release-monitoring.org also appears to have notifications for new
versions, which might be a better alternative to this script, but my
thinking was it might be good to periodically generate a full report
of how out of date we are.

I am unsure as of yet how many of our core-plans are covered by
release-monitoring, but if there is interest in this approach, I don't
mind slowly going through the list and finding all the project IDs.

Signed-off-by: Steven Danna <steve@chef.io>