Add script to find packages we forgot to promote
In addition to merging the relevant PRs to this repository, it is
important that we promote the build group produced by the
merge. However, sometimes we forgot to do this when the build take a
while or when the build group will promote something that we are
unsure about.

This is a small script maintainers can use to periodically get us back
up to date.

Signed-off-by: Steven Danna <steve@chef.io>