[ruby] add quick tests for Ruby plans
Quick smoke check to make sure Ruby built, is the expected version, and
provides some common Ruby commands to do dep-solving.

Each Ruby plan gets a `tests/` directory and a script in each that
sources the one reuseable test.sh script for Ruby builds.

In the "real" test script ruby/tests/test.sh, BASH_SOURCE[0] is used for
referencing this actual script's directory.  It is better in this case
than $0 which would reference the calling location and would mess up the
execution of the single, reuseable BATS script.

Signed-off-by: Robb Kidd <robb@thekidds.org>