[nss-myhostname] add a new plan!
nss-myhostname[1] is a plugin for the GNU Name Service Switch (NSS)
functionality of the GNU C Library (glibc) providing host name
resolution for the locally configured system hostname as returned by
gethostname(2). This plugin is commonly enabled in the default
`/etc/nsswitch.conf` NSS configuration of RHEL and RHEL-like Linux
distributions, with a hosts entry like:

    hosts:      files dns myhostname

glibc has `/etc/nsswitch.conf` hard-coded as the file location for
configuration.  So long as core/glibc ships with that setting, this
plugin is necessary for some software to perform name resolution with
expected behavior.

An example of unexpected behavior: Ruby.

It is common in the Ruby ecosystem to check if a domain name is
resolvable by using the Ruby stdlib Socket class to attempt to open a
network socket to a domain and port. (44,402 results searching for "rescue
Socket" in Ruby code[2] at the time of this writing.) With core/ruby* as
built today, performing on any RHEL-flavored host a minimal name
resolution check on a domain expected to fail returns Errno::EBUSY.

    $ hab pkg exec core/ruby ruby -r socket -e "p Socket.pack_sockaddr_in(80, 'nope.example')"
    Traceback (most recent call last):
            1: from -e:1:in `<main>'
    -e:1:in `pack_sockaddr_in': Device or resource busy - getaddrinfo (Errno::EBUSY)

getaddrinfo finds itself in an else-branch returning Errno::EBUSY[3]
because it was unable to find lookup functions whose names match the
plugins listed in `/etc/nsswitch.conf`.

I was unable to find any examples of Ruby code handling Errno::EBUSY
exceptions when a Socket or SocketError was expected.

A Ruby built with a core/nss-myhostname as a runtime dependency succeeds
(or rather, fails in the expected manner).

    $ hab pkg exec core/ruby ruby -r socket -e "p Socket.pack_sockaddr_in(80, 'nope.example')"
    Traceback (most recent call last):
            1: from -e:1:in `<main>'
    -e:1:in `pack_sockaddr_in': getaddrinfo: No address associated with hostname (SocketError)

* Why kernel2 builds?

My specific interest in expected name resolution failures in Ruby is in
service to packaging Ruby projects (e.g. Chef Infra Client and Chef
InSpec) in Habitat for use on a wide variety of platforms and platform
versions. For those projects to work on RHEL-flavored hosts still
running Linux v2, they need a kernel2 build of this plugin.

This project is not expected to itself generate much churn in publishing
new packages. This current version was released on May 9, 2011.

[1] http://0pointer.de/lennart/projects/nss-myhostname/
[2] https://github.com/search?l=Ruby&q=%22rescue+Socket%22&type=Code
[3] https://github.com/bminor/glibc/blob/glibc-2.27/sysdeps/posix/getaddrinfo.c#L926-L939

Signed-off-by: Robb Kidd <robb@thekidds.org>