[binutils] Re-enable new-dtags as the default
As of binutils 2.29, the patch we had to enable new-dtags as the
default no longer works. Because of this, we've been building
executables with RPATH rather than RUNPATH since late 2017
(6f67b5cfb).

The ld.so manual page explains how RUNPATH and RPATH are used as
follows:

>   -  Using the directories specified in the DT_RPATH dynamic section
>      attribute of the binary if present and DT_RUNPATH attribute does
>      not exist.  Use of DT_RPATH is deprecated.
>
>   -  Using the environment variable LD_LIBRARY_PATH, unless the
>      executable is being run in secure-execution mode (see below), in
>      which case this variable is ignored.
>
>   -  Using the directories specified in the DT_RUNPATH dynamic section
>      attribute of the binary if present.  Such directories are searched
>      only to find those objects required by DT_NEEDED (direct
>      dependencies) entries and do not apply to those objects' children,
>      which must themselves have their own DT_RUNPATH entries.  This is
>      unlike DT_RPATH, which is applied to searches for all children in
>      the dependency tree.
>
>   -  From the cache file /etc/ld.so.cache...
>
>   -  In the default path /lib, and then /usr/lib....

(http://man7.org/linux/man-pages/man8/ld.so.8.html)

This PR removes the patch and replaces it with a configure option that
has the intended affect.

Not that because this has been broken for some considerable time,
fixing this may introduces behave changes for users that they aren't
expected.

Fixes #3333

Signed-off-by: Steven Danna <steve@chef.io>