[Docs] Add guide document for base plans
> As a maintainer of core-plans I want to understand why some plans are considered
'base plans, and why only some of those can be updated outside of a refresh,
and why some need to wait.

> As a user of core plans, I want to know what the base plans refresh is, why
it is necessary and what it entails.

This PR attempts to explain at a high level what a base plan is and why they require
extra care when merging. It also breaks apart which plans require a refresh to be
updated, and why, and which can be updated outside of a refresh.

Finally, the process around the refresh, why it exists and how it might change as
Builder is able to handle more complex workloads is covered.

Signed-off-by: Scott Macfarlane <smacfarlane@chef.io>