[runc] Update to runc 1.0.0-rc10
In addition to updating to the current version of `runc`, this commit
addresses a handful of problems with the `runc` plan.

First, despite a `pkg_version` being specified in the plan, the build
logic from the Go scaffolding resulted in unspecified code being
built. You would end up building whatever happened to be on the master
branch of https://github.com/opencontainers/runc at the time of your
build.

Now, the download logic has been overridden to grab the code on the
branch specified by `pkg_version` so we can a) reliably build the same
thing over and over, and b) have `pkg_version` correspond to the
contents of the resulting package.

Second, we now create a static binary.

Third, the installation logic of the Go scaffolding has been
overridden. Previously, despite running `make` in the `do_build`
callback, we would run `go install` in the `do_install` callback. This
ended up running `go build` and then installing *that* binary in our
package. Since there is considerably more logic in `runc`'s Makefile,
this actually resulted in a binary *without* compiled support for
`libseccomp`, despite it being a build dependency! Now, we simply copy
the binary created from the `do_build` callback, and everything works.

(The `pkg_shasum` line was removed because it wasn't being used by
anything.)

Tests for all these conditions are included in `tests/test.bats`.

Signed-off-by: Christopher Maier <cmaier@chef.io>