[dex] bump v2.24.0 -> v2.25.0, simplify plan.sh
This is now modelled after opa/plan.sh, downloading the tarball, working with
that.

Since 2.25.0, dex no longer vendors its dependencies, so there's no need to
tender any GOPATH.

To have `dex version` return its version correctly, this fiddles with one of the
scripts.  However, if it means we don't have to mess with `go install`, `go
build`, and the proper ldflags, but can use the make target instead, I think
that's preferable.

Signed-off-by: Stephan Renatus <srenatus@chef.io>